Bounceportal can now be specified
The portal/recall command you wish to use for roomshere noportal is
flagged can now be set. It does not yet save or store the data in the
tables yet because I'm not sure what that needs to look like and I don't
have a working copy of serialize for Blowtorch.