Noted added about rawcolor
This version 0.2.1 uses the gmcp setting "rawcolor on". This breaks any
gmcp comms plugins as they will now have color codes ie @x012 or @b all
over the place but it allows the database to search for names which
previously had ansi color codes hidden in them and would return nothing
when a room, stored in the databse, had color codes in it's name.